+++ android/src/main/java/com/vitanov/multiimagepicker/MultiImagePickerFileProvider.java
@@ -1,7 +1,0 @@
-package com.vitanov.multiimagepicker;
-
-import androidx.core.content.FileProvider;
-
-public class MultiImagePickerFileProvider extends FileProvider {
-
-}