+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+### Changed (v4.2.2)
+
+## 2019-05-08
+
+- Added option to customize the message when max selection limit is reached on Android
+
 ### Changed (v4.2.1)
 
 ## 2019-05-03